Fix some compilation warnings and spamming runtime logged warnings in rd-cpp
* Fix some compilation warnings:
  * Add explicit casts from size_t to int32_t
  * Ignore MSVC warnings 4996 and 4250
  * Proper wstring->string conversion
* Make error message one-lined
* Remove spamming warning "No need to explicitly terminate lifetime definition in C++"
* Remove copy-pasted logging message from kotlin rd implementation

+++ rd-cpp/src/rd_framework_cpp/src/main/wire/PkgInputStream.cpp
@@ -41,7 +41,7 @@ int32_t PkgInputStream::try_read(Buffer::word_t* res, size_t size)
 			return -1;
 		}
 	}
-	const int n = (std::min)(size, memory - buffer.get_position());
+	const int32_t n = static_cast<int32_t>((std::min)(size, memory - buffer.get_position()));
 	Buffer::word_t* start = buffer.current_pointer();
 	std::copy(start, start + n, res);
 	buffer.set_position(buffer.get_position() + n);
